src/requirements.txt
lxml==4.9.2
openai==0.27.8
PyYAML==6.0
tqdm==4.65.0
wandb==0.15.4
torch==2.0.1
scipy==1.10.1
bitsandbytes==0.39.0
evaluate==0.4.0
sacrebleu==2.3.1
datasets==2.13.0
Xformers==0.0.20
datasets==2.13.0
transformers @ git+https://github.com/huggingface/transformers.git
peft @ git+https://github.com/huggingface/peft.git
accelerate @ git+https://github.com/huggingface/accelerate.git
